Make the h2 register lock same as mysql.
